Refactor Message and Post Controllers to Use AuthenticatedUserService
- Updated MessageController to remove direct user ID handling and utilize AuthenticatedUserService for user retrieval.
- Modified createMessage method to validate and set the authenticated user.
- Adjusted PostController to implement AuthenticatedUserService for user association during post creation.
- Enhanced error handling in PostController for better response on creation failure.

Refactor Token and User Controllers for Improved Readability

- Cleaned up imports in TokenController and UserController for better readability.
- Removed unused imports and organized existing ones.

Update DTOs for Nullable Fields

- Changed id fields in ForumResponseDTO and MessageResponseDTO to be nullable.
- Updated constructor signatures in DTOs to reflect nullable types.

Enhance Entity Relationships

- Added utilisateur relationship in Post entity to associate posts with users.
- Updated Utilisateur entity to manage posts collection, including add/remove methods.

Implement AuthenticatedUserService

- Created AuthenticatedUserService to handle user authentication logic.
- Integrated service into MessageController and PostController for user retrieval.

Fix Frontend Issues

- Updated CreateMachine.vue to correctly remove bullets from the form.
- Adjusted Discussion.vue to remove unused user ID from message posting.
- Fixed Forum.vue to ensure correct API endpoint usage for forum creation.
- Modified MaterielInfo.vue to fetch machine data correctly and handle loading states.
- Updated Posts.vue to correct API endpoint for fetching posts related to a forum.

General Code Cleanup

- Removed unused imports across various files for improved code clarity.
- Standardized code formatting and organization in multiple files.

+++ back/.php-cs-fixer.dist.php
@@ -18,7 +18,7 @@ return (new Config())
         'phpdoc_to_comment' => true,
         'explicit_string_variable' => true,
         'concat_space' => ['spacing' => 'one'],
-        'array_syntax' => ['syntax' => 'long'],
+        'array_syntax' => ['syntax' => 'short'],
     ])
     ->setFinder($finder)
 ;
